requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
audioread==3.0.0
cachetools==5.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
cmake==3.26.3
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
einops==0.6.0
fastdtw==0.3.4
ffmpeg-python==0.2.0
filelock==3.8.0
fonttools==4.38.0
future==0.18.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.11.0
HyperPyYAML==1.2.1
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
jiwer==2.5.1
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.20.2
librosa==0.10.0.post2
lit==16.0.5
llvmlite==0.40.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
more-itertools==9.0.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
noisereduce==1.1.0
num2words==0.5.12
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper==20230314
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pooch==1.6.0
praat-textgrids==1.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
rapidfuzz==2.13.2
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.9.3
sentencepiece==0.1.99
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
speechbrain==0.5.14
sympy==1.12
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tiktoken==0.3.1
tokenizers==0.13.2
torch==2.0.1
torchaudio==2.0.2
tqdm==4.64.1
transformers==4.24.0
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.11.0
